implements gfr-simple test case
and also adds structure to how mapped commits will work. basically sometimes we will need to handle when something says from X, but we didnt include X, so we need to see what our latest known ancestor of X is, so we maintain a map of those marks. to complicate things a bit, the relationship can be transitive, ie we want to merge from X, but we dont have X, but X merged to Y, but we dont have Y, but Y merged to Z and we did have Z, so we say X -> Z.