dont try to preserve empty merge commits
originally i wanted full compatibility with git-filter-repo in regards to how it keeps some merge commits that become empty. i looked at how it filtered some complex histories like react or linux, and it introduces a lot of empty merge commits from a result of pruning. i decided thats not really desirable, so i changed the requirement that we explicitly try to avoid including those empty merge commits by default. as a result the tests can be simplified a bit because ultimately all we care about is: does our current HEAD ls-tree match the ls-tree of the git-filter-repo HEAD? if so, then the number of commits doesnt matter because we are guaranteed that we are at the same point in time. Additionally i added back the react repo test case because this is a good example of a simple test where we just check that our ls-tree matches git-filter-repo ls-tree, and not worry about the actual commits (we do count the number of non-empty commits, and it does match, so that means we are missing N commits where all N are empty commits so it shouldnt matter anyway