callback uses ShouldAddMode response, adds ConsecutiveCommitGroups
response can be Add, DontAdd, Exit, or AddAndExit. if its one of the exit ones, then it returns that to the caller and the caller can decide what to do based on if it should exit or not. in our case, the default implementation will call child.kill(). ConsecutiveCommitGroups is a simple wrapper struct to track the indices of which commits are going to be returned to the user, and return them in consecutive groups instead of a flat list of random indices