fixes child.wait() never closing on large streams
the issue was that the generate_commit_list_from_lines fn would error if it found non-utf8 line, and then the line_reader would never close, which would cause the child.wait() to not close as well. so 2 solutions here: 1. if there is any error in parsing, we should force kill the child. 2. replace the line reader with a manual reader that uses utf8 lossy because apparently there can be git messages that have non-utf8 chars, and that throws off the stream