re-implements topbase alg using simpler method
this seperate function only cares about topbasing, not fullbasing/rewind. in the future I will add another seperate function for rewind mode and that will be useful for diff-log, but fullbase might not be necessary so might get rid of that. This new topbase function returns a guarantee of a fork point whereas the last function was overly complex and its return value was implicit