IMPORTANT: changes how mgt topbase considers a fork point. see note
Previously, when looking for fork points, we only cared if the top branch's commit's blobs all existed somewhere in the base branch. This is not correct because you can find a commit X that has two blobs Q, and U, and those blob can be found in potentially two different commits in the base branch, but wed treat that as 'existing' in the base branch. Now, we changed our topbase algorithm to not just look at the blob id, but rather a hashed version of all of the information in the blob summary given by 'git log --raw'. So we now hash the file modes, the src and dest SHA's and the path. This commit changes the test case because the old test case would not be valid under this new topbase algorithm because its comparing file1 being created against file1 being modified, which would have a different hash. Maybe in the future there will be different hashing modes which will allow this old test case to work, but I think that this new way is the most correct because it tells us exactly if this blob exists in the other branch without any ambiguity