implements try_sync_in
a couple notes here: its important to note that our rebase interactive is a bit wonky here. we are implicitly relying on the fact that after we do an actual gitfilter that the output commits on the filtered branch will be the same as the calculated commits we did when doing the in-memory topbase. This assumption might not hold... Now that earlier i realized its possible to do a gitfilter and a rebase without having to switch branches, maybe it would be better to run gitfilter first on the temporary branch, and then collect the topbase information from that, such that when we tell the user if they want to push/pull, the commits that we are displaying to them can be the exact ones that will be part of the rebase. Otherwise what happens now is that the user sees the commits from the pre-filtered branch, which can be confusing.