solves merge commit TODO by not allowing merge commits to be synced
because most of the time they cannot be. A merge commit when filtered would just be an empty commit anyway. Also it is confusing for the user if they get a list of commits saying 'heres what we can push/pull' but if theres one dangling merge commit on either side, that would confuse them, and also fail if they tried to push/pull that single merge commit. just dont show them merge commits, and dont allow them when rebasing/filtering