fixes logic flow of handling cleanup. see note:
previoulsly, it was failing to delete the temp branch on cleanup. The issue was that it was returning early as soon as it had an error. this solution right now is very messy. TODO: clean up all of the try_* functions such that they can proceed to the next recovery step even if the previous ones fail