fixes improper merge commit filtering, see note:
adds better file contents checking (hashing the file ops, and checking if commit A's file ops are a subset of commit B's file ops set. Adds better ancestry detection without recursively traversing a graph. Adds temporary debug logging. IMPORTANT: reverses merge commit ordering explicitly. Something is wrong with how im recreating the vec of merge marks because its in the wrong order. Here i added a .reverse() to explicitly reverse that and make it 'right', but ideally it would be in the right order in the first place