topbase can now use merge commits as fork points, see note:
Adds Debug trait bound for hashing mode, which is fine because they all implement that now anyway. Removes incorrect logic of filtering out merge commits in the actual topbase algorithm (you can optionally filter them out at a higher level if desired). Changes how BlobHashingMode::EndState detects state (it now sets everything except Deleted and Renamed to be treated as Modified) because we want an Add 0 -> X to be treated the same as a Modified Z -> X, because end state only cares about the end state, not where it came from. Topbase function now uses BlobHashingMode::EndState because thats the only way to adequately test if a merge commit is a fork point