removes unimplemented tests for now, see note:
The last test in this file is particularly difficult to implement. basically we want to be able to pull in the latest updates from a remote repo, and if the remote repo has a merge commit with a conflict resolution, that resolution should exist when we rebase it/filter it onto our local repo. The problem is that even with --rebase-merges in git rebase, git will not reapply that conflict resolution automatically. Furthermore, --rebase-merges requires us to parse a lot more information to be able to properly structure the labels and merges in the interactive todo text. So for now, we will not have the ability to sync merge commits